refactor: move grid to its own module
