+++ src/constants.js
@@ -1,0 +1,8 @@
+export const GRID_SIZE = 20;
+export const LINE_DEFAULT_WIDTH = 8
+export const HIGHLIGHTED_LINE_DEFAULT_WIDTH = LINE_DEFAULT_WIDTH + 2
+export const WORLD_WIDTH = 1000;
+export const WORLD_HEIGHT = 1000;
+export const GHOST_POINT_RADIUS = 5;
+export const GHOST_POINT_COLOR = 0x00ff00;
+export const DEFAULT_HIGHLIGHT_OPTIONS = {glow: false}
